so much progress and performance improvements on the system.
reworked a bunch and made everything so much nicer to deal with. Should have just used a big array (also easier to 'inspect') yesterday or from the start and saved myself 24 real-time hours lol.

+++ Assets/Scripts/GameManager.cs
@@ -15,7 +15,7 @@ namespace GMTK
 	public class GameManager : TimelineListener
 	{
 		public static Action<GameState> OnGameStateChange;
-		public long Score => _timeline.CurrentDisplayedFrame;
+		public int Score => _timeline.CurrentDisplayedFrame;
 
 		public GameState GameState => _gameState;
 		//only serialized for testing to see it in the inspector.
@@ -35,7 +35,7 @@ namespace GMTK
 			OnGameStateChange?.Invoke(_gameState);
 		}
 		
-		private void OnInput(long frame, GameInput input, bool instant)
+		private void OnInput(int frame, GameInput input, bool instant)
 		{
 			if (frame <= 1)
 			{
